Add files via upload
This file includes code for the line sensor, and some default motion execution of the chassis. It also includes a list of nodes for the first default root (follow the path from start position to junction between the two factories). Code for the line sensor (Frank) was tested and passed. Code for the default root (Fiana) was tested but failed & need further adjustments.